moved waiter and keepalivewatcher into a concurrency package
