+++ lib/app/src/main/java/claw/CLAWRobot.java
@@ -7,8 +7,8 @@ import java.io.StringWriter;
 import claw.logs.CLAWLogger;
 import claw.logs.LogHandler;
 import claw.rct.commands.CommandLineInterpreter;
-import claw.rct.network.low.Waiter;
-import claw.rct.network.low.Waiter.NoValueReceivedException;
+import claw.rct.network.low.concurrency.Waiter;
+import claw.rct.network.low.concurrency.Waiter.NoValueReceivedException;
 import claw.rct.remote.RCTServer;
 import edu.wpi.first.wpilibj.Preferences;
 import edu.wpi.first.wpilibj.TimedRobot;
