+++ lib/app/src/main/java/claw/rct/network/low/concurrency/Waiter.java
@@ -1,4 +1,4 @@
-package claw.rct.network.low;
+package claw.rct.network.low.concurrency;
 
 import java.util.Optional;
